backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.2
django-cors-headers==4.3.1
django-filter==24.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
python-decouple==3.8
python-dateutil==2.8.2
Pillow==10.1.0
